Disable checkbox for sending emails if not admin (fixes #9)

+++ src/components/Modals/DeleteDocumentModal.tsx
@@ -63,7 +63,7 @@ export const DeleteDocumentModal = ({ projectPk, documentPk, documentKind, refet
                     refetchProgressYears(() => {
                         reset();
                     });
-                    onDeleteSuccess();
+                    onDeleteSuccess && onDeleteSuccess();
 
                 }
                 // onClose();
